requirements.txt
Django==1.7.4
argparse==1.2.1
python-dateutil==2.4.2
pytz==2014.10
six==1.9.0
wsgiref==0.1.2
requests==2.6.0
django-tastypie==0.12.1

